fix: panic caused by `set_hook`
