+++ yazi-fm/src/app/commands/mod.rs
@@ -1,2 +1,3 @@
 mod plugin;
+mod render;
 mod stop;
